fix(self-lint): lint workspace members and clear the backlog
`just self-lint` ran `cargo dylint --all -- --all-targets` with no
`--workspace`, so the args after `--` defaulted `cargo check` to the
root package plus its dependency graph. The standalone `tools/*` crates
are not dependencies of root, so their own-lint violations accumulated
undetected while `just all` stayed green.

Add `--workspace` to the recipe (CI delegates to `just self-lint`, so
no separate CI edit is needed) and clear the now-surfaced backlog:

- auto-fixable: `prefer_raw_string`, `macro_trailing_comma`,
  `import_grouping`, `import_granularity`, `unicode_ellipsis_in_docs`,
  `unicode_ellipsis_in_comments`.
- `single_letter_closure_param` / `single_letter_generic`: rename the
  flagged parameters.
- `unit_test_file_layout`: extract the over-budget inline `#[cfg(test)]
  mod tests` blocks in `tools/gen-docs` into external `tests.rs` files.
- `macro_argument_binding`: ignore syn's `Token!` in `dylint.toml` —
  its token-literal argument has no runtime expression to bind.
