+++ tools/gen-docs/src/render/markdown.rs
@@ -8,12 +8,10 @@
 use std::sync::LazyLock;
 
 use pulldown_cmark::{CodeBlockKind, CowStr, Event, Options, Tag, TagEnd, html as cmark_html};
-use syntect::{
-    highlighting::{Theme, ThemeSet},
-    html::{ClassStyle, ClassedHTMLGenerator, css_for_theme_with_class_style},
-    parsing::SyntaxSet,
-    util::LinesWithEndings,
-};
+use syntect::highlighting::{Theme, ThemeSet};
+use syntect::html::{ClassStyle, ClassedHTMLGenerator, css_for_theme_with_class_style};
+use syntect::parsing::SyntaxSet;
+use syntect::util::LinesWithEndings;
 
 pub(crate) fn markdown_to_html(markdown: &str) -> String {
     let mut options = Options::empty();
@@ -44,7 +42,7 @@ fn highlight_code_blocks<'a>(parser: impl Iterator<Item = Event<'a>>) -> Vec<Eve
                     .split(',')
                     .next()
                     .map(str::trim)
-                    .filter(|s| !s.is_empty())
+                    .filter(|part| !part.is_empty())
                     .map(str::to_owned);
                 if let Some(lang) = lang {
                     current_lang = Some(lang);
@@ -94,18 +92,18 @@ fn highlight_to_html(code: &str, lang: &str) -> String {
     format!("<pre><code{class_attr}>{body}</code></pre>\n")
 }
 
-/// Render the `language-…` class attribute for a code block, but
+/// Render the `language-...` class attribute for a code block, but
 /// only when the fence's language tag is made of characters that
 /// are unambiguously safe to drop into an HTML attribute. Anything
 /// outside the safe-character set yields an empty attribute — the page
 /// still renders correctly because the inline `<span class>`es
 /// generated by `syntect` carry the actual highlighting, and the
-/// `language-…` class is purely decorative here.
+/// `language-...` class is purely decorative here.
 fn lang_class_attr(lang: &str) -> String {
     let safe = !lang.is_empty()
         && lang
             .chars()
-            .all(|c| c.is_ascii_alphanumeric() || matches!(c, '_' | '-' | '+' | '#' | '.'));
+            .all(|ch| ch.is_ascii_alphanumeric() || matches!(ch, '_' | '-' | '+' | '#' | '.'));
     if safe {
         format!(" class=\"language-{lang}\"")
     } else {
@@ -132,14 +130,14 @@ pub(crate) static HIGHLIGHT_CSS: LazyLock<String> = LazyLock::new(|| {
 });
 
 /// Render a short, single-line snippet of markdown without the
-/// outer `<p>…</p>` wrapper that block-level rendering inserts. Used
+/// outer `<p>...</p>` wrapper that block-level rendering inserts. Used
 /// for table cells and inline headings where backticks should still
 /// turn into `<code>`.
 pub(crate) fn markdown_inline_to_html(markdown: &str) -> String {
     let parser = pulldown_cmark::Parser::new(markdown).filter(|event| {
         !matches!(
             event,
-            Event::Start(Tag::Paragraph) | Event::End(TagEnd::Paragraph)
+            Event::Start(Tag::Paragraph) | Event::End(TagEnd::Paragraph),
         )
     });
     let mut buffer = String::new();
